+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/_internal/state/HubdleState.kt
@@ -465,6 +465,7 @@ internal data class HubdleState(
                 }
 
                 data class Features(
+                    var compose: Boolean = false,
                     var coroutines: Boolean = false,
                     var extendedStdlib: Boolean = true,
                     var extendedTesting: Boolean = true,